renaming files (switched point and line)
